+++ src/lib/wsman-xml-serialize.h
@@ -58,7 +58,7 @@ struct __XmlSerializationData
 //	char* name; // optional can be NULL
 //	char* nameNs;
 	int mode;
-	int index;
+	unsigned int index;
 	WsXmlNodeH xmlNode;
     XML_NODE_ATTR *attrs;
 //	int indentCount;
@@ -97,11 +97,11 @@ void enforce_mustunderstand_if_needed(WsContextH cntx, WsXmlNodeH node);
 
 //int do_serialize_uint(struct __XmlSerializationData* data, int valSize);
 
-void* ws_serializer_alloc(WsContextH cntx, int size);
+void* ws_serializer_alloc(WsContextH cntx, size_t size);
 //int do_serializer_free(WsContextH cntx, void* ptr);
 int ws_serializer_free(WsContextH cntx, void* ptr);
 
-void *xml_serializer_alloc(XmlSerializationData *data, int size, int zeroInit);
+void *xml_serializer_alloc(XmlSerializationData *data, size_t size, int zeroInit);
 
 int xml_serializer_free(XmlSerializationData *data, void *buf);
